z_hook: Reroll currentlevel

+++ dplay/src/functions.cpp
@@ -5,4 +5,4 @@ BYTE* (__fastcall * LoadFileInMem)(const char* pszName) = reinterpret_cast<BYTE*
 void (__fastcall * AddL1Objs)(int, int, int, int) = reinterpret_cast<void (__fastcall *)(int, int, int, int)>(0x00459015);
 void (__fastcall * AddMazeBook)(int, int, int, int, int, int, int, int, int) = reinterpret_cast<void (__fastcall *)(int, int, int, int, int, int, int, int, int)>(0x00458D47);
 void (__fastcall * SpawnUnique)(int, int, int) = reinterpret_cast<void (__fastcall *)(int, int, int)>(0x00421CDC);
-void (__fastcall * NetSendCmd)(int, int, int, int, int) = reinterpret_cast<void (__fastcall *)(int, int, int, int, int)>(0x0046ED1E);
+void (__fastcall * NetSendCmd)(int, BYTE, BYTE, int, int) = reinterpret_cast<void (__fastcall *)(int, BYTE, BYTE, int, int)>(0x0046ED1E);
